allow decimal for price

+++ app/controllers/editProfileController.js
@@ -279,7 +279,7 @@ exports.setPrice = function(requestBody, type, callback) {
     console.log("Query: " + query)
     dbController.queryGet(query, (result) => {
         if(result.status == 200) {
-            if (parseInt(result.body.rows[0].base_price) > parseInt(requestBody.price)) {
+            if (parseFloat(result.body.rows[0].base_price) > parseFloat(requestBody.price)) {
                 callback([], "Price should be higher than base price!");
             } else {
                 const query1 = `UPDATE can_care SET price = ${requestBody.price} WHERE username = '${user}' AND type = '${type}';`;
